Add 2 minute stale time to the more filters selectors
react-query sets the default stale time to 0, and there are many
circunstances where this is undesirable.

refetchOnMount is true by default, as well as refetchOnWindowFocus and
so. This means that everytime the more filters menu opened three calls
to the server were made. By adding a 2 minute stale time we limit this
calls