refactor(IndicatorValues): refactors code to allow for easier treatment of new Indicators
Refactors code on places related to Indicator value calculation (such as creating an scenario intervention when creating new IndicatorRecords, or querying for a RiskMap), so that it'0s not hardcoded and can dynamically do calculations depending on the indicator that is being queried. Once new Indicators are added, they need to be implemented in the code as new strategies. Indicator's nameCode has also been made non-null to avoid possible issues
Sourcing Data Import hasn't been refactored yet, because of potential performance issues.