Use autoUpdate instead of manual update()
The select positions have to change many times: if the selection
changes the container might become higher, so the menu needs to be
displaced. If there's not enough space and the position switches to the
top, filtering the elements might leave the menu floating in the air.

These cases where handled with useLayoutEffect and detecting them
manually, but we're switching to a more robust way, just allowing the
library to update them as neccessary.