Add require-await ESLint rule
A lot of calls to useQuery were made with an async function that
internally used promise chaining. This means that the top level function
was resolved immediatly.

In practice, this meant that `isLoading` was always false, `isSuccess`
always true... the only thing stopping us from accessing undefined data
is the placeholder data.

I'm not sure of why typescript didn't complain, but it looks like
react-query can handle that case, as it populated the data correctly
anyway when the fetch succeeded.